Refactored string definitions into seperate file
